new headers in common
non-functional state
