+++ src/fringe/dll_version.h
@@ -1,4 +1,3 @@
-#include <charconv>
 namespace LinkedMap {
 
 struct Link {
